Give geofencing its own subdirectory
