+++ event-details/arrival-tracking/geofencing/index.ts
@@ -1,0 +1,2 @@
+export * from "./Geofencer"
+export * from "./ExpoGeofencer"
